split gpg-interface into individual source files
