+++ include/builtin.h
@@ -11,8 +11,7 @@
  * 		int cmd_<builtin>(int argc, char *argv[])
  *
  * - Add extern function prototype to builtin.h
- * - In main.c, add builtin to builtins[]:
- *
+ * - Register builtin by adding new member of registered_builtins struct:
  * 		{ "<builtin>", cmd_<builtin> },
  * */
 
@@ -28,5 +27,18 @@ extern int cmd_init(int argc, char *argv[]);
 extern int cmd_message(int argc, char *argv[]);
 extern int cmd_publish(int argc, char *argv[]);
 extern int cmd_read(int argc, char *argv[]);
+extern int cmd_import_key(int argc, char *argv[]);
+
+struct cmd_builtin registered_builtins[] = {
+		{ "channel", cmd_channel },
+		{ "config", cmd_config },
+		{ "init", cmd_init },
+		{ "message", cmd_message },
+		{ "publish", cmd_publish },
+		{ "get", cmd_get },
+		{ "read", cmd_read },
+		{ "import-key", cmd_import_key },
+		{ NULL, NULL }
+};
 
 #endif //GIT_CHAT_BUILTIN_H
