+++ include/fs-utils.h
@@ -1,6 +1,8 @@
 #ifndef GIT_CHAT_FS_UTILS_H
 #define GIT_CHAT_FS_UTILS_H
 
+#include <stddef.h>
+
 #include "strbuf.h"
 
 /**
