+++ include/str-array.h
@@ -1,7 +1,7 @@
 #ifndef GIT_CHAT_STR_ARRAY_H
 #define GIT_CHAT_STR_ARRAY_H
 
-#include <stdlib.h>
+#include <stddef.h>
 #include <stdarg.h>
 
 /**
