+++ src/gnupg/gpg-common.c
@@ -1,0 +1,98 @@
+#include <fcntl.h>
+#include <errno.h>
+#include <locale.h>
+
+#include "gpg-common.h"
+#include "working-tree.h"
+#include "fs-utils.h"
+
+const char *get_gpgme_library_version()
+{
+	return gpgme_check_version(NULL);
+}
+
+NORETURN void GPG_FATAL(const char *msg, gpgme_error_t err)
+{
+	FATAL("%s: %s: %s", msg, gpgme_strsource(err), gpgme_strerror(err));
+}
+
+void init_gpgme_openpgp_engine(void)
+{
+	gpgme_error_t err;
+	int errsv = errno;
+
+	LOG_INFO("Initializing GPGME with GPGME_PROTOCOL_OpenPGP engine");
+
+	const char *version = gpgme_check_version(NULL);
+	LOG_INFO("Using installed GPGME version %s", version);
+
+	setlocale(LC_ALL, "");
+	err = gpgme_set_locale(NULL, LC_CTYPE, setlocale(LC_CTYPE, NULL));
+	if (err)
+		GPG_FATAL("failed to set GPGME locale", err);
+
+#ifdef LC_MESSAGES
+	err = gpgme_set_locale (NULL, LC_MESSAGES, setlocale (LC_MESSAGES, NULL));
+	if (err)
+		GPG_FATAL("failed to set GPGME locale", err);
+#endif
+
+	err = gpgme_engine_check_version(GPGME_PROTOCOL_OpenPGP);
+	if (err && gpgme_err_code(err) != GPG_ERR_NO_ERROR)
+		GPG_FATAL("failed to load OpenPGP engine using installed version of GPGME", err);
+
+	struct _gpgme_engine_info *enginfo;
+	err = gpgme_get_engine_info(&enginfo);
+	if (err)
+		GPG_FATAL("failed to load gpgme engine info", err);
+
+	const char *engine_version = enginfo->version;
+	const char *protocol_name = (char *) gpgme_get_protocol_name(GPGME_PROTOCOL_OpenPGP);
+	const char *engine_executable = enginfo->file_name;
+	const char *engine_config_dir = enginfo->home_dir;
+	const char *default_config_dir = gpgme_get_dirinfo("homedir");
+
+	if (!engine_version)
+		engine_version = "cannot be determined";
+	if (!protocol_name)
+		protocol_name = "invalid protocol number";
+	if (!engine_executable)
+		engine_executable = "unknown";
+	if (!engine_config_dir)
+		engine_config_dir = default_config_dir;
+	if (!engine_config_dir)
+		engine_config_dir = "unknown";
+
+	LOG_DEBUG("gpgme crypto engine version number: %s", engine_version);
+	LOG_DEBUG("gpgme engine protocol: %s", protocol_name);
+	LOG_DEBUG("gpgme protocol implementation binary: %s", engine_executable);
+	LOG_DEBUG("gpgme config directory: %s", engine_config_dir);
+
+	errno = errsv;
+}
+
+void gpgme_context_init(struct gc_gpgme_ctx *ctx, int use_gc_gpg_homedir)
+{
+	struct strbuf gpg_homedir;
+	strbuf_init(&gpg_homedir);
+	if (get_gpg_homedir(&gpg_homedir))
+		FATAL("local GPG home directory does not exist or cannot be used for some reason");
+
+	gpgme_error_t err = gpgme_new(&ctx->gpgme_ctx);
+	if (err)
+		GPG_FATAL("failed to create GPGME context", err);
+
+	if (use_gc_gpg_homedir)
+		safe_create_dir(gpg_homedir.buff, NULL, S_IRUSR | S_IWUSR | S_IXUSR);
+
+	err = gpgme_ctx_set_engine_info(ctx->gpgme_ctx, GPGME_PROTOCOL_OpenPGP, NULL,
+			use_gc_gpg_homedir ? gpg_homedir.buff : NULL);
+	if (err)
+		GPG_FATAL("failed to mutate GPGME engine homedir", err);
+}
+
+void gpg_context_release(struct gc_gpgme_ctx *ctx)
+{
+	gpgme_release(ctx->gpgme_ctx);
+	strbuf_release(&ctx->gnupg_homedir);
+}
