+++ src/str-array.c
@@ -1,3 +1,4 @@
+#include <stdlib.h>
 #include <string.h>
 
 #include "str-array.h"
